+++ sdk/digitaltwins/azure-digitaltwins-core/src/main/java/com/azure/digitaltwins/core/DigitalTwinsServiceVersion.java
@@ -17,7 +17,12 @@ public enum DigitalTwinsServiceVersion implements ServiceVersion {
     /**
      * Service version {@code 2021-06-30-preview}.
      */
-    V2021_06_30_PREVIEW("2021-06-30-preview");
+    V2021_06_30_PREVIEW("2021-06-30-preview"),
+    
+    /**
+     * Service version {@code 2022-05-31}.
+     */
+    V2022_05_31("2022-05-31");
 
     private final String version;
 
@@ -38,6 +43,6 @@ public enum DigitalTwinsServiceVersion implements ServiceVersion {
      * @return The latest service API version of Azure Digital Twins that is supported by this client.
      */
     public static DigitalTwinsServiceVersion getLatest() {
-        return V2021_06_30_PREVIEW;
+        return V2022_05_31;
     }
 }