+++ sdk/digitaltwins/azure-digitaltwins-core/CHANGELOG.md
@@ -1,9 +1,16 @@
 # Release History
 
-## 1.3.0-beta.1 (2022-03-25)
+## 2.0.0 (2022-06-30)
 
 ### Features Added
+- Updated service API version to use API version 2022-05-31 by default.
 - Added `sourceTime` to `DigitalTwinPropertyMetadata`.
+- Added a new class `DigitalTwinComponentMetadata` for deserializing component metadata.
+- Added support for `lastUpdatedOn` for digital twins and components to `DigitalTwinMetadata` and `DigitalTwinComponentMetadata`.
+
+### Fixes and improvements
+
+- Updated samples to illustrate usage of `sourceTime`, `lastUpdatedOn` at digital twin and component metadata level, and use of `DigitalTwinComponentMetadata`.
 
 ## 1.2.0 (2022-03-11)